+++ gse-app/src/main/resources/lang/ProjectPane.properties
@@ -3,9 +3,6 @@ CreateFolder=Create folder
 Data=Data
 Delete=Delete
 DoYouConfirm=Do you confirm?
-DragError=Drag error
-DragFileExists=The dragged file already exists
-Error=Error
 FileWillBeDeleted=File %s will be deleted
 FilesWillBeDeleted=Files %s will be deleted
 Name=Name
