+++ gse-app/src/main/resources/lang/ProjectPane_fr.properties
@@ -3,9 +3,6 @@ CreateFolder=Créer dossier
 Data=Données
 Delete=Supprimer
 DoYouConfirm=Confirmez-vous?
-DragError=Erreur de déplacement
-DragFileExists=Le fichier déplacé existe déja
-Error=Erreur
 FileWillBeDeleted=Le fichier %s sera supprimé
 FilesWillBeDeleted=Les fichiers %s seront supprimés
 Name=Nom
