+++ gse-util/src/main/java/com/powsybl/gse/util/DragAndDropMove.java
@@ -6,12 +6,24 @@
  */
 package com.powsybl.gse.util;
 
-import javafx.scene.control.TreeItem;
+import com.powsybl.afs.*;
+import javafx.scene.control.*;
+import javafx.scene.input.*;
+import javafx.scene.paint.Color;
+
+import java.util.ResourceBundle;
 
 public class DragAndDropMove {
-    private Object source;
-    private TreeItem sourceTreeItem;
 
+    private static final ResourceBundle RESOURCE_BUNDLE = ResourceBundle.getBundle("lang.DragAndDropMove");
+
+    private boolean nameFound;
+
+    private boolean success;
+
+    private Object source;
+
+    private TreeItem sourceTreeItem;
 
     public Object getSource() {
         return source;
@@ -28,4 +40,197 @@ public class DragAndDropMove {
     public void setSourceTreeItem(TreeItem sourceTreeItem) {
         this.sourceTreeItem = sourceTreeItem;
     }
+
+    public boolean isMovable(Object item, TreeItem targetTreeItem) {
+        return (item instanceof ProjectFolder || item instanceof Folder) && item != getSource() && targetTreeItem != getSourceTreeItem().getParent();
+    }
+
+    public void dragDetectedEvent(Object value, TreeItem treeItem, MouseEvent event, TreeView treeView) {
+        setSource(value);
+        setSourceTreeItem(treeItem);
+        if (value instanceof ProjectNode && treeItem != treeView.getRoot()) {
+            Dragboard db = treeView.startDragAndDrop(TransferMode.ANY);
+            ClipboardContent cb = new ClipboardContent();
+            cb.putString(((ProjectNode) value).getName());
+            db.setContent(cb);
+            event.consume();
+        }
+    }
+
+    public void dragDetectedEvent(Object value, TreeItem treeItem, MouseEvent event, TreeTableView tree) {
+        setSource(value);
+        setSourceTreeItem(treeItem);
+        if (value instanceof Project && treeItem != tree.getRoot()) {
+            Dragboard db = tree.startDragAndDrop(TransferMode.ANY);
+            ClipboardContent cb = new ClipboardContent();
+            cb.putString(((Project) value).getName());
+            db.setContent(cb);
+            event.consume();
+        }
+    }
+
+    public void dragOverEvent(DragEvent event, Object item, TreeItem targetTreeItem, TreeCell treeCell) {
+        if (isMovable(item, targetTreeItem) && !isSourceAncestorOf(targetTreeItem)) {
+            boolean nameSearch = false;
+            treeItemChildrenSize(targetTreeItem, nameSearch);
+            textFillColor(treeCell);
+            event.acceptTransferModes(TransferMode.ANY);
+            event.consume();
+        }
+    }
+
+    public void dragOverEvent(DragEvent event, Object item, TreeTableRow treeTableRow, TreeTableCell treetableCell) {
+        if (isMovable(item, treeTableRow.getTreeItem()) && !isSourceAncestorOf(treeTableRow.getTreeItem())) {
+            boolean nameSearch = false;
+            treeItemChildrenSize(treeTableRow.getTreeItem(), nameSearch);
+            textFillColor(treetableCell);
+            event.acceptTransferModes(TransferMode.ANY);
+            event.consume();
+        }
+    }
+
+    public void dragDroppedEvent(Object value, TreeItem treeItem, DragEvent event, ProjectNode projectNode) {
+        if (value instanceof ProjectFolder && value != getSource()) {
+            ProjectFolder projectFolder = (ProjectFolder) projectNode;
+            boolean search = false;
+            success = false;
+            treeItemChildrenSize(treeItem, search);
+            accepTransferDrag(projectFolder, success);
+            event.setDropCompleted(success);
+            refresh(getSourceTreeItem().getParent());
+            refresh(treeItem);
+            event.consume();
+        }
+    }
+
+    public void dragDroppedEvent(Object value, TreeItem treeItem, DragEvent event, Node node) {
+        if (value instanceof Folder && value != getSource()) {
+            Folder folder = (Folder) node;
+            boolean search = false;
+            success = false;
+            treeItemChildrenSize(treeItem, search);
+            accepTransferDrag(folder, success);
+            event.setDropCompleted(success);
+            refresh(getSourceTreeItem().getParent());
+            refresh(treeItem);
+            event.consume();
+        }
+    }
+
+    private void textFillColor(TreeCell treeCell) {
+        if (!nameExists()) {
+            treeCell.setTextFill(Color.CHOCOLATE);
+        }
+    }
+
+    private void textFillColor(TreeTableCell treeTableCell) {
+        if (!nameExists()) {
+            treeTableCell.setTextFill(Color.CHOCOLATE);
+        }
+    }
+
+
+    private boolean nameExists() {
+        return nameFound;
+    }
+
+
+    private boolean isSourceAncestorOf(TreeItem targetTreeItem) {
+        TreeItem treeItemParent = targetTreeItem.getParent();
+        while (treeItemParent != null) {
+            if (getSourceTreeItem() == treeItemParent) {
+                return true;
+            } else {
+                treeItemParent = treeItemParent.getParent();
+            }
+
+        }
+        return false;
+    }
+
+    private void treeItemChildrenSize(TreeItem treeItem, boolean nameMatch) {
+        nameFound = nameMatch;
+        if (!treeItem.isLeaf()) {
+            if (treeItem.getValue() instanceof ProjectFolder) {
+                projectFolderChildrenSize(treeItem);
+            } else if (treeItem.getValue() instanceof Folder) {
+                folderChildrenSize(treeItem);
+            }
+        }
+    }
+
+    private void projectFolderChildrenSize(TreeItem treeItem) {
+        ProjectFolder treeItemFolder = (ProjectFolder) treeItem.getValue();
+        if (!treeItemFolder.getChildren().isEmpty()) {
+            for (ProjectNode node : treeItemFolder.getChildren()) {
+                if (node == null) {
+                    break;
+                } else if (node.getName().equals(getSource().toString())) {
+                    nameFound = true;
+                }
+            }
+        }
+    }
+
+    private void folderChildrenSize(TreeItem treeItem) {
+        Folder folder = (Folder) treeItem.getValue();
+        if (!folder.getChildren().isEmpty()) {
+            for (Node node : folder.getChildren()) {
+                if (node == null) {
+                    break;
+                } else if (node.getName().equals(getSource().toString())) {
+                    nameFound = true;
+                }
+            }
+        }
+    }
+
+    /**
+     * @param projectFolder
+     * @param s
+     */
+    private void accepTransferDrag(ProjectFolder projectFolder, boolean s) {
+        success = s;
+        if (nameExists()) {
+            nameAlreadyExistsAlert();
+        } else {
+            ProjectNode monfichier = (ProjectNode) getSource();
+            monfichier.moveTo(projectFolder);
+            success = true;
+        }
+    }
+
+    /**
+     * @param folder
+     * @param s
+     */
+    private void accepTransferDrag(Folder folder, boolean s) {
+        success = s;
+        if (nameExists()) {
+            nameAlreadyExistsAlert();
+        } else {
+            Project monfichier = (Project) getSource();
+            monfichier.moveTo(folder);
+            success = true;
+        }
+    }
+
+
+    private Alert nameAlreadyExistsAlert() {
+        Alert alert = new Alert(Alert.AlertType.ERROR);
+        alert.setTitle(RESOURCE_BUNDLE.getString("DragError"));
+        alert.setHeaderText(RESOURCE_BUNDLE.getString("Error"));
+        alert.setContentText(RESOURCE_BUNDLE.getString("DragFileExists"));
+        alert.showAndWait();
+        return alert;
+    }
+
+    private void refresh(TreeItem item) {
+        if (item.getValue() instanceof ProjectFolder || item.getValue() instanceof Folder) {
+            item.setExpanded(false);
+            item.setExpanded(true);
+        }
+    }
+
+
 }
