+++ gse-util/src/main/resources/lang/DragAndDropMove.properties
@@ -1,0 +1,3 @@
+DragError=Drag error
+DragFileExists=The dragged file already exists
+Error=Error
