+++ gse-util/src/main/resources/lang/DragAndDropMove_fr.properties
@@ -1,0 +1,3 @@
+DragError=Erreur de déplacement
+DragFileExists=Le fichier déplacé existe déja
+Error=Erreur
