+++ gse-util/src/main/resources/lang/NodeChooser.properties
@@ -1,13 +1,8 @@
 ConfirmationDialog=Confirmation dialog
 CreateFolder=Create folder
 Delete=Delete
-DeleteError=Delete error
-DeletedFolderContainsProjects=The folder %s contains at least one project
 Description=Description
 DoYouConfirm=Do you confirm?
-DragError=Drag error
-DragFileExists=The dragged file already exists
-Error=Error
 File=File
 FileWillBeDeleted=File %s will be deleted
 FilesWillBeDeleted=Files %s will be deleted
