+++ gse-util/src/main/resources/lang/NodeChooser_fr.properties
@@ -1,12 +1,8 @@
 ConfirmationDialog=Fenètre de confirmation
 CreateFolder=Créer un dossier
 Delete=Supprimer
-DeleteError=Erreur de Suppression
-DeletedFolderContainsProjects=le Dossier %s Contient au moins un projet
 Description=Description
 DoYouConfirm=Confirmez vous?
-DragError=Erreur de déplacement
-DragFileExists=le fichier déplacé existe déja
 Error=Erreur
 File=Fichier
 FileWillBeDeleted=Le fichier %s sera supprimé
